+++ CHANGELOG.md
@@ -2,16 +2,21 @@
 
 ## [Unreleased]
 
+- [TD-2486] Template type `domain`
+
+## [4.9.0] 2020-11-30
+
 ### Added
 
 - [TD-3089] Widget and type `copy` on df
-- [TD-2486] Template type `domain`
+- [TD-3066] Deprecate implementations and publish audit event if a referenced
+  structure is deleted
 
 ## [4.8.0] 2020-11-16
 
 ### Added
 
-- [TD-2092] Download implementations in csv format
+- [TD-2092] Download implementations in CSV format
 
 ## [4.7.0] 2020-11-03